requirements.txt
azure-ai-projects
azure-identity
python-dotenv
gradio
fastapi
uvicorn
flask
requests
matplotlib
azure.ai.ml